[REFACTOR]: Logic Projects
